# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZeroRisco API — transporte seguro em Saquarema, RJ
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Cadastrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                password:
                  type: string
                  minLength: 6
                role:
                  type: string
                  enum: [passenger, driver]
              required: [name, email, phone, password]
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: Login com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Renovar access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: Tokens renovados
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Alterar senha
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
              required: [currentPassword, newPassword]
      responses:
        "200":
          description: Senha alterada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /rides/history:
    get:
      operationId: getRidesHistory
      tags: [rides]
      summary: Histórico de corridas do passageiro (paginação por cursor)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          description: Cursor ISO-8601 da última corrida retornada (campo createdAt)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Lista paginada de corridas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRides"
  /rides/driver/history:
    get:
      operationId: getDriverRidesHistory
      tags: [rides]
      summary: Histórico de corridas do motorista (paginação por cursor)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Lista paginada de corridas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDriverRides"
  /rides/driver/stats:
    get:
      operationId: getDriverStats
      tags: [rides]
      summary: Estatísticas de ganhos e corridas do motorista
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Estatísticas do motorista
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverStats"
  /rides/promo/validate:
    post:
      operationId: validatePromoCode
      tags: [rides]
      summary: Validar código promocional
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required: [code]
      responses:
        "200":
          description: Código válido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoValidation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /rides/promo/active:
    get:
      operationId: getActivePromoCodes
      tags: [rides]
      summary: Listar cupons ativos (visíveis ao passageiro)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de cupons ativos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoCodePublic"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-secret
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflito — recurso já existe
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit atingido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
            phone:
              type: string
            isApproved:
              type: boolean
            rgStatus:
              type: string
            profilePhotoUrl:
              type: string
              nullable: true
...
    PaginatedRides:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RideHistoryItem"
        nextCursor:
          type: string
          nullable: true
          description: Passe como ?cursor= na próxima requisição para buscar mais
        hasMore:
          type: boolean
      required: [items, hasMore]
    PaginatedDriverRides:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DriverRideHistoryItem"
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
      required: [items, hasMore]
    RideHistoryItem:
      type: object
      properties:
        id:
          type: string
        originAddress:
          type: string
        destAddress:
          type: string
        rideType:
          type: string
        price:
          type: number
        status:
          type: string
        distance:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        waitTimeFee:
          type: number
        promoCode:
          type: string
          nullable: true
        promoDiscount:
          type: number
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        driver:
          nullable: true
          type: object
          properties:
            name:
              type: string
            rating:
              type: number
            vehicleModel:
              type: string
              nullable: true
            vehiclePlate:
              type: string
              nullable: true
            vehicleColor:
              type: string
              nullable: true
    DriverRideHistoryItem:
      type: object
      properties:
        id:
          type: string
        originAddress:
          type: string
        destAddress:
          type: string
        rideType:
          type: string
        price:
          type: number
        status:
          type: string
        distance:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        driverRating:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    DriverStats:
      type: object
      properties:
        totalRides:
          type: integer
        totalEarnings:
          type: number
        weekRides:
          type: integer
        weekEarnings:
          type: number
        monthRides:
          type: integer
        monthEarnings:
          type: number
        todayRides:
          type: integer
        rating:
          type: number
    PromoValidation:
      type: object
      properties:
        valid:
          type: boolean
        code:
          type: string
        description:
          type: string
          nullable: true
        discountType:
          type: string
        discountValue:
          type: number
    PromoCodePublic:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
          nullable: true
        discountType:
          type: string
        discountValue:
          type: number
        expiresAt:
          type: string
          format: date-time
          nullable: true
...